requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
elevenlabs
google-generativeai
reportlab
faster-whisper
